refactor : 바우처와 회원의 Creators 클래스 위치 변경
