+++ src/test/java/kr/co/springbootweeklymission/member/application/MemberServiceTest.java
@@ -1,0 +1,4 @@
+package kr.co.springbootweeklymission.member.application;
+
+public class MemberServiceTest {
+}
