+++ src/test/java/kr/co/springbootweeklymission/voucher/domain/repository/JdbcVoucherRepositoryTest.java
@@ -2,10 +2,10 @@ package kr.co.springbootweeklymission.voucher.domain.repository;
 
 import kr.co.springbootweeklymission.infrastructure.error.exception.NotFoundException;
 import kr.co.springbootweeklymission.infrastructure.error.model.ResponseStatus;
-import kr.co.springbootweeklymission.member.domain.creators.MemberCreators;
+import kr.co.springbootweeklymission.member.creators.MemberCreators;
 import kr.co.springbootweeklymission.member.domain.entity.Member;
 import kr.co.springbootweeklymission.member.domain.repository.MemberRepository;
-import kr.co.springbootweeklymission.voucher.domain.creators.VoucherCreators;
+import kr.co.springbootweeklymission.voucher.creators.VoucherCreators;
 import kr.co.springbootweeklymission.voucher.domain.entity.Voucher;
 import kr.co.springbootweeklymission.voucher.domain.model.VoucherPolicy;
 import org.junit.jupiter.api.MethodOrderer;
